fix: moved skills and inv to respective utils
