+++ utils/databaseUtils.js
@@ -151,159 +151,6 @@ exports.getNewLevelRewards = async function(id, level, channel) {
     channel.send("Vous avez atteint le niveau " + level + " !");
 }
 
-/**
- * Adding an item to the inventory of a player or increasing the quantity of an existing item
- * @param playerId id of the player
- * @param {*} item item to give
- * @param {*} quantity quantity of the item to give
- */
-exports.giveItem = async function(playerId, item, quantity) {
-    const playerCollection = Client.mongoDB.db('player-data').collection(playerId);
-
-    // Querying the inventory in the database
-    const inventory = await playerCollection.findOne(
-        {name: "inventory"}, 
-        {projection: {_id: 0, skills: 0, activeSkills: 0}}
-    );
-    
-    // If the item is already in the inventory, we add the quantity to the existing one
-    // Else, we create a new entry for the item
-    if(inventory.items[item] != undefined) {
-        inventory.items[item].quantity += quantity;
-    } else {
-        inventory.items = {
-            ...inventory.items,
-            [item]: {
-                quantity: quantity
-            }
-        }
-    }
-
-    // Updating the inventory in the database
-    const result = await playerCollection.updateOne({name: "inventory"}, { $set: { items: inventory.items } }, { upsert: true });
-
-    console.group("[DEBUG] Item " + item + " given to player " + playerId);
-}
-
-exports.learnSkill = async function(id, skill_id)  {
-    const playerCollection = Client.mongoDB.db('player-data').collection(id);
-
-    const query = { name: "inventory" };
-    let options = { 
-        projection: {_id: 0, items: 0, quantity: 0},
-    };
-
-    const inventory = await playerCollection.findOne(query, options);
-    
-    if(inventory.skills.includes(skill_id)) {
-        console.log("[DEBUG] Skill " + skill_id + " already exists for user " + id + ".");
-        return false;
-    } else {
-        const newItems = inventory.skills.concat(skill_id);
-        const update = { $set: { skills: newItems } };
-        options = { upsert: true };
-        const result = await playerCollection.updateOne(query, update, options);
-        console.log("[DEBUG] User " + id + " learned skill " + skill_id + ".");
-        return true;
-    }
-}
-
-exports.unlearnSkill = async function(user_id, skill_id) {
-    const playerCollection = Client.mongoDB.db('player-data').collection(user_id);
-
-    const query = { name: "inventory" };
-    let options = { 
-        projection: {_id: 0, items: 0, quantity: 0},
-    };
-
-    const inventory = await playerCollection.findOne(query, options);
-    
-    if(inventory.skills.includes(skill_id)) {
-        const newItems = [...(inventory.skills)];
-        const removeIndex = newItems.indexOf(skill_id);
-        if(removeIndex > -1)
-            newItems.splice(removeIndex, 1);
-        const update = { $set: { skills: newItems } };
-        options = { upsert: true };
-        const result = await playerCollection.updateOne(query, update, options);
-        console.log("[DEBUG] User " + user_id + " forgot skill " + skill_id + ".");
-    } else {
-        console.error("[DEBUG] Skill " + skill_id + " hasn't been learned by user " + user_id + ". (UNLEARN_SKILL_MISSING)");
-    }
-}
-
-exports.hasSkill = async function(user_id, skill_id) {
-    const playerCollection = Client.mongoDB.db('player-data').collection(user_id);
-
-    const query = { name: "inventory" };
-    let options = { 
-        projection: {_id: 0, items: 0, quantity: 0},
-    };
-
-    const inventory = await playerCollection.findOne(query, options);
-    
-    if(inventory.skills.includes(skill_id)) {
-        return true;
-    } else  {
-        return false;
-    }
-}
-
-exports.selectActiveSkill = async function(user_id, skill_id) {
-    const playerCollection = Client.mongoDB.db('player-data').collection(user_id);
-
-    const query = { name: "inventory" };
-    let options = { 
-        projection: {_id: 0, items: 0, quantity: 0},
-    };
-
-    const inventory = await playerCollection.findOne(query, options);
-    
-    if(!inventory.skills.includes(skill_id)) {
-        console.error("[DEBUG] Skill " + skill_id + " hasn't been learned by user " + user_id + ". (ACTIVE_SKILL_NOT_LEARNED)");
-        return false;
-    }
-    if(inventory.activeSkills.length > 4) {
-        console.error("[DEBUG] 4 Active Skills already used by user " + user_id + ". (ACTIVE_SKILL_MAX_SIZE)");
-        return false;
-    }
-    if(inventory.activeSkills.includes(skill_id)) {
-        console.error("[DEBUG] Skill " + skill_id + " already selected active by " + user_id + ". (ACTIVE_SKILL_ALREADY_ACTIVE)");
-        return false;
-    }
-
-    const newItems = inventory.activeSkills.concat(skill_id);
-    const update = { $set: { activeSkills: newItems } };
-    options = { upsert: true };
-    const result = await playerCollection.updateOne(query, update, options);
-    console.log("[DEBUG] User " + user_id + " selected skill " + skill_id + ".");
-    return true;
-}
-
-exports.unselectActiveSkill = async function(user_id, skill_id) {
-    const playerCollection = Client.mongoDB.db('player-data').collection(user_id);
-
-    const query = { name: "inventory" };
-    let options = { 
-        projection: {_id: 0, items: 0, quantity: 0},
-    };
-
-    const inventory = await playerCollection.findOne(query, options);
-    
-    if(inventory.activeSkills.includes(skill_id)) {
-        const newItems = [...(inventory.skills)];
-        const removeIndex = newItems.indexOf(skill_id);
-        if(removeIndex > -1)
-            newItems.splice(removeIndex, 1);
-        const update = { $set: { activeSkills: newItems } };
-        options = { upsert: true };
-        const result = await playerCollection.updateOne(query, update, options);
-        console.log("[DEBUG] User " + user_id + " unselected skill " + skill_id + ".");
-    } else {
-        console.error("[DEBUG] Skill " + skill_id + " hasn't been selected by user " + user_id + ". (ACTIVE_SKILL_NOT_ACTIVE)");
-    }
-}
-
 
 exports.setClass = async function(id, className) {
     const playerCollection = Client.mongoDB.db('player-data').collection(id);
